// src/App.types.ts
export enum UrlEnum {
  cpmMap = '/cpm-map',
  schedule = '/schedule',
  createOrder = '/create-order',
  listProvodka = '/list-provodka',
  analytics = '/analytics',
  changeForm = '/change-form',
  tableProvodka = '/table-provodka',
  page1 = '/page1',
  page2 = '/page2',
  sandbox = '/sandbox',
}